requirements.txt
numpy
matplotlib
scikit-learn
Pillow
imagecodecs
rasterio
torch
torchvision
tqdm